+++ api/src/database/loaders/items/ItemsLoader.ts
@@ -1,5 +1,10 @@
+import Item from "@Database/entities/items/Item";
 import LogsManager from "@Logs";
 import SHabbo from "@SHabbo";
+import { DeepPartial } from "typeorm";
+import * as ItemsData from "./ItemsData.json";
+
+type ItemPartial = DeepPartial<Item>;
 
 class ItemsLoader implements ILoader {
     async beforeRun(): Promise<boolean> {
@@ -8,12 +13,11 @@ class ItemsLoader implements ILoader {
     }
 
     async run(): Promise<void> {
-        // Mode
-        await SHabbo.getDatabase().getItems().create({ id: 152, name: 'table_polyfon', width: 2, length: 2, height: 1 }).save();
-        await SHabbo.getDatabase().getItems().create({ id: 25, name: 'stand_polyfon_z', height: 1 }).save();
-        await SHabbo.getDatabase().getItems().create({ id: 31, name: 'table_polyfon_med', width: 2, length: 2, height: 1 }).save();
-        await SHabbo.getDatabase().getItems().create({ id: 17, name: 'table_polyfon_small', width: 2, length: 2, height: 1 }).save();
-        await SHabbo.getDatabase().getItems().create({ id: 18, name: 'chair_polyfon', height: 1, can_sit: true }).save();
+        const items: ItemPartial[] = ItemsData;
+
+        for await (const item of items) {
+            await SHabbo.getDatabase().getItems().create(item).save();
+        }
 
         LogsManager.success('Default items inserted.');
     }